back-end/crms/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MongoDB Driver (Motor for async support)
motor==3.3.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# JWT & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client for service communication
httpx==0.26.0

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
requests